Refactoring for CORS config
